+++ src/main/java/svnserver/server/command/LockManyCmd.java
@@ -106,4 +106,11 @@ public final class LockManyCmd extends BaseCmd<LockManyCmd.Params> {
         .listEnd()
         .listEnd();
   }
+
+  @Override
+  protected void permissionCheck(@NotNull SessionContext context, @NotNull Params args) throws IOException, SVNException {
+    for (PathRev pathRev : args.paths) {
+      context.checkWrite(context.getRepositoryPath(pathRev.path));
+    }
+  }
 }